Took the code that was in Program.cs and moved it into a utility class. Cleaned up the code, unwrapping nesting and some variable names changes and such. No logic should be changed.
